Pull shared ec2/elb connection code into AWS common code
